add checkbox to risks and activity
