+++ packages/frontend/src/pages/scaling-risk/props/getRiskView.ts
@@ -31,6 +31,7 @@ export function getRiskViewEntry(
     showProjectUnderReview: isAnySectionUnderReview(project),
     isVerified,
     isUpcoming: project.isUpcoming,
+    technology: project.display.category,
     ...project.riskView,
   }
 }
