+++ packages/frontend/src/pages/scaling-risk/view/ScalingRiskView.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import { ActiveIcon } from '../../../components/icons/symbols/ActiveIcon'
 import { ArchivedIcon } from '../../../components/icons/symbols/ArchivedIcon'
 import { ScalingLegend } from '../../../components/ScalingLegend'
+import { RollupsOnlyCheckbox } from '../../../components/table/filters/checkboxes/RollupsOnlyCheckbox'
 import { getScalingRowProps } from '../../../components/table/props/getScalingRowProps'
 import { getScalingRiskColumns } from '../../../components/table/props/getScalingTableColumns'
 import { RowConfig, TableView } from '../../../components/table/TableView'
@@ -22,6 +23,7 @@ export function ScalingRiskView({ items }: ScalingRiskViewProps) {
 
   return (
     <section className="mt-4 sm:mt-8">
+      <RollupsOnlyCheckbox className="mb-4" />
       <Tabs
         items={[
           {
